+++ sh/common_safe.sh
@@ -9,7 +9,7 @@ if [[ ${rpc_url:-unset} = 'unset' ]] ; then
     rpc_url="$(get_api_secret rpcUrl)"
     declare -r rpc_url
 fi
-if [[ -z $rpc_url ]] ; then
+if [[ ${rpc_url:-unset} = 'unset' ]] ; then
     echo '`rpcUrl` is unset in `api_secrets.json` for chain "'"$chain_name"'"' >&2
     exit 1
 fi